Move UnsafeRich* to common scalanative.unsiged package for cross-compilation gurantees
